# back_end/swagger/full_api.yml
swagger: "2.0"
info:
  title: SOP Management API
  version: "1.0"
basePath: /api
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json

paths:
  /get_sop:
    get:
      summary: Get all sop names
      description: Returns a list of SOP names
      responses:
        200:
          description: List of SOP names
          schema:
            type: array
            items:
              type: string
          examples:
            application/json: ["SA6118", "SA6293"]

  /run_code:
    post:
      summary: Api for run SOP
      description: Get all Step from Db --> run
      parameters:
        - in: body
          name: dataObj
          required: true
          schema:
            $ref: '#/definitions/getAllStepObject'
      responses:
        200:
          description: "Get All Step successfull and run it"
        
  /get_action:
    get:
      summary: Get all action name
      description: Returns a list of Action names
      responses:
        200:
          description: List of Action names
          schema:
            type: array
            items:
              type: string
          examples:
              application.json: ["open", "click"]

  /add_action:
    post:
      summary: add action name
      parameters:
        - in: body
          name: dataObj
          required: true
          schema:
            type: object
            required: 
              - actionName
            properties:
              actionName:
                type: string
                example: 'open'
      responses:
        200:
          description: Add new action successfull

  /delete_sop:
    delete:
      summary: Delete SOP
      description: Delete all data relate to the SOP
      parameters:
      - in: body
        name: dataObj
        required: true
        schema:
          type: object
          required:
            - sop_name
          properties:
            sop_name:
              type: string
              example: 'SA6118'
      responses:
        200:
          description: Name of SOP were deleted

  /add_sop:
    post:
      summary: Add new SOP
      description: Add new SOP to DB by SOP name
      parameters:
      - in: body
        name: dataObj
        required: true
        schema:
          type: object
          required: 
            - sop_name
          properties:
            sop_name:
              type: string
              example: 'SA6118'
      responses:
        200:
          description: Add new SOP successfull
  
  /add_step:
    post:
      summary: Add Step into an SOP.
      description: add new step, increase the step order by existing one.
      parameters:
        - in: body
          name: dataObj
          required: true
          schema:
            $ref: '#/definitions/stepObject'
      responses:
        200:
          description: Add Step successfull

  /update_step_value:
    post:
      summary: Update each step key value if needed
      description: Update step by stepId
      parameters:
      - in: body
        name: dataObj
        required: true
        schema:
          $ref: '#/definitions/updateStepObject'
      responses:
        200:
          description: Update Step

  /get_all_step:
    post:
      summary: Get all step from an SOP
      description: Get all Step by sop_name and sop_type
      parameters:
        - in: body
          name: dataObj
          required: true
          schema:
            $ref: '#/definitions/getAllStepObject'
      responses:
        200:
          description: Get all step successfull

  /delete_step_order:
    delete:
      summary: Delete specific Step
      description: Delete Step by stepId
      parameters:
        - in: body
          name: dataObj
          required: true
          schema:
            type: object
            required:
              - stepOrder
              - setUpId
            properties:
              stepOrder:
                  type: integer
                  example: 1
              setUpId:
                  type: integer
                  example: 1
      responses:
        200:
          description: Delete step successfull


definitions:
  stepObject:
    type: object
    required:
      - sop_name
      - setUpType
      - actionName
      - xPath
      - text
      - folder_path
      - file_name
    properties:
      sop_name:
        type: string
        example: 'SA6117'
      setUpType:
        type: string
        example: 'initial'
      actionName:
        type: string
        example: 'click'
      xPath:
        type: string
        example: "//div[text()='']"
      text:
        type: string
        example: 'admin'
      folder_path:
        type: string
      file_name:
        type: string
  updateStepObject:
    type: object
    required: 
      - stepId
      - xPath
      - text
      - folder_path
      - file_name
      - actionName
    properties:
      stepId:
        type: integer
      xPath:
        type: string
        example: "//div123"
      actionName:
        type: string
        example: "click"
      text:
        type: string
        example: "password"
      folder_path:
        type: string
        example: "C://"
      file_name:
        type: string
        example: "SwConfigFile"
  getAllStepObject:
    type: object
    required:
      - sop_name
      - setup_type
    properties:
      sop_name: 
        type: string
        example: "SA6118"
      setup_type: 
        type: string
        example: "initial"